Fix warnings for compares with differing signedness

+++ src/utils.cpp
@@ -152,14 +152,14 @@ void LA_AVDECC_CALL_CONVENTION displayAssertDialog(char const* const file, unsig
 	{
 		char buffer[2048];
 		auto offset = std::snprintf(buffer, sizeof(buffer), "Debug Assertion Failed!\n\nFile: %s\nLine: %u\n\n\t", file, line);
-		if (offset < sizeof(buffer) - 1)
+		if (offset < static_cast<ssize_t>(sizeof(buffer) - 1))
 		{
 			offset += std::vsnprintf(buffer + offset, sizeof(buffer) - offset, message, arg);
 			buffer[sizeof(buffer) - 1] = 0; // Contrary to std::snprintf, std::vsnprintf does not add \0 if there is not enough room
 
 			std::cerr << buffer << std::endl;
 
-			if (offset < sizeof(buffer) - 1)
+			if (offset < static_cast<ssize_t>(sizeof(buffer) - 1))
 			{
 				offset += std::snprintf(buffer + offset, sizeof(buffer) - offset,
 					"\n"
